refactor(BaseballGame): divide into two separate classes
Added new class called BaseballGame which:
- generate an answer for baseball game
- validate user's input for answer
- parse String of user's input
- calculate strikes and balls of user's guess

Removed corresponding functions in Application class
